Added the GPSService Service and an example activity
GPSService is a service that used the location manager and NOT the the
better fusedlocation YET. Might happen later. Know problem are random
spikes in location need lowpass?

Also added an example activty on how to use it. It has an PERMISSION
check implemented that is universal and can be used in ANY code.
:dancer: