... Forgot to add permission check when entering the New track activity.
That should do it! Now we check in the main GUI that we have
persmission, if not we deny the user access to that activity and asks it
politly yet strict as fuq to give them all their honor and respect that
they may have left so that our program can swim in the glory of people
who lost their basic permissions. ¨hue256